improvments

+++ domain/beers.go
@@ -9,11 +9,11 @@ type Beer struct {
 }
 
 type BeerType struct {
-	ID   int    `json:"id"`
+	ID   int64  `json:"id"`
 	Name string `json:"name"`
 }
 
 type BeerStyle struct {
-	ID   int    `json:"id"`
+	ID   int64  `json:"id"`
 	Name string `json:"name"`
 }
